+++ src/main/java/com/oronaminc/join/websocket/session/CustomWebSocketHandlerDecorator.java
@@ -1,4 +1,4 @@
-package com.oronaminc.join.websocket.config;
+package com.oronaminc.join.websocket.session;
 
 import java.security.Principal;
 import java.util.Objects;
